+++ packages/paste-core/components/input/src/Input.tsx
@@ -45,7 +45,7 @@ export const InputElement = styled.input<InputProps>((props) =>
     display: 'block',
     fontFamily: 'inherit',
     fontSize: 'fontSize30',
-    fontWeight: 'fontWeightNormal',
+    fontWeight: 'fontWeightMedium',
     lineHeight: 'lineHeight20',
     outline: 'none',
     paddingBottom: 'space30',
@@ -67,6 +67,9 @@ export const InputElement = styled.input<InputProps>((props) =>
     '&:disabled': {
       color: props.variant === 'inverse' ? 'colorTextInverseWeaker' : 'colorTextWeaker',
       cursor: 'not-allowed',
+      // Fixes value color in Safari
+      '-webkit-text-fill-color': props.variant === 'inverse' ? 'colorTextInverseWeaker' : 'colorTextWeaker',
+      '-webkit-opacity': '1',
     },
   })
 );