+++ packages/paste-style-props/src/types/typography.ts
@@ -1,14 +1,14 @@
 // https://styled-system.com/api/#typography
-import * as CSS from 'csstype';
-import {ThemeShape} from '@twilio-paste/theme';
-import {ResponsiveValue, TLengthStyledSystem} from '@twilio-paste/styling-library';
+import type * as CSS from 'csstype';
+import type {ThemeShape} from '@twilio-paste/theme';
+import type {ResponsiveValue, TLengthStyledSystem} from '@twilio-paste/styling-library';
 
 // Tokens
 export type FontFamilyOptions = keyof ThemeShape['fonts'] | 'inherit';
 export type FontSizeOptions = keyof ThemeShape['fontSizes'] | '100%' | 'inherit';
 export type FontWeightOptions = keyof ThemeShape['fontWeights'] | 'inherit';
 export type LineHeightOptions = keyof ThemeShape['lineHeights'] | 'unset' | 'inherit';
-export type TextColorOptions = keyof ThemeShape['textColors'] | 'currentColor' | 'inherit';
+export type TextColorOptions = keyof ThemeShape['textColors'] | 'currentColor' | 'inherit' | 'transparent';
 
 export type FontFamily = ResponsiveValue<FontFamilyOptions>;
 export type FontSize = ResponsiveValue<FontSizeOptions>;
@@ -23,6 +23,7 @@ export type FontStyleOptions = CSS.FontStyleProperty;
 export type WhiteSpaceOptions = CSS.WhiteSpaceProperty;
 export type WordBreakOptions = CSS.WordBreakProperty;
 export type WordWrapOptions = CSS.WordWrapProperty;
+export type OverflowWrapOptions = CSS.OverflowWrapProperty;
 export type TextDecorationOptions = CSS.TextDecorationProperty<CSS.TextDecorationLineProperty>;
 export type TextOverflowOptions = CSS.TextOverflowProperty;
 export type TextTransformOptions = CSS.TextTransformProperty;
@@ -33,6 +34,7 @@ export type FontStyle = ResponsiveValue<FontStyleOptions>;
 export type WhiteSpace = ResponsiveValue<WhiteSpaceOptions>;
 export type WordBreak = ResponsiveValue<WordBreakOptions>;
 export type WordWrap = ResponsiveValue<WordWrapOptions>;
+export type OverflowWrap = ResponsiveValue<OverflowWrapOptions>;
 export type TextDecoration = ResponsiveValue<TextDecorationOptions>;
 export type TextOverflow = ResponsiveValue<TextOverflowOptions>;
 export type TextTransform = ResponsiveValue<TextTransformOptions>;
@@ -58,4 +60,5 @@ export interface TypographyProps {
   whiteSpace?: WhiteSpace;
   wordBreak?: WordBreak;
   wordWrap?: WordWrap;
+  overflowWrap?: OverflowWrap;
 }