+++ packages/paste-theme/src/themes/index.ts
@@ -1,5 +1,6 @@
 import {DefaultTheme} from './default';
 import {SendGridTheme} from './sendgrid';
 import {ConsoleTheme} from './console';
+import {DarkTheme} from './dark';
 
-export {DefaultTheme, ConsoleTheme, SendGridTheme};
+export {DefaultTheme, ConsoleTheme, SendGridTheme, DarkTheme};